+++ crates/utils/src/async.rs
@@ -2,5 +2,6 @@ pub use futures::future::FusedFuture;
 pub use futures::future::join_all;
 pub use futures::future::try_join_all;
 pub use futures::select;
+pub use futures::select_biased;
 pub use futures::stream;
 pub use futures::stream::StreamExt;